requirements.txt
librosa==0.11.0
numpy>=1.24.0
matplotlib>=3.5.0
pyacoustid==1.3.0 